pdf-chatbot/backend/requirements.txt
# PDF Processing
camelot-py==0.11.0
pymupdf==1.23.8
pandas==1.5.3
tabulate==0.9.0
opencv-python==4.8.1.78

# LlamaIndex
llama-index==0.9.48

# Vector Database
qdrant-client==1.7.0

# LLM
openai>=1.40.0
azure-identity==1.15.0

# Reranking
sentence-transformers>=2.3.0

# BM25
rank-bm25==0.2.2

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
typing-extensions==4.8.0
